added builder and dockerized builder

+++ builder/builder.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/sh
 
 set -e
 
@@ -13,20 +13,23 @@ GITHUB_REPO="$1"
 DOCKER_REPO="$2"
 CLONE_DIR="temp-repo"
 
-read -p "Docker Hub Username: " DOCKER_USERNAME
-read -s -p "Docker Hub Password (or access token): " DOCKER_PASSWORD
-echo ""
+# Check environment variables
+if [ -z "$DOCKER_USER" ] || [ -z "$DOCKER_PWD" ]; then
+  echo "Error: DOCKER_USER and DOCKER_PWD environment variables must be set."
+  exit 1
+fi
 
 echo "Cloning $GITHUB_REPO..."
 rm -rf "$CLONE_DIR"
 git clone "$GITHUB_PREF$GITHUB_REPO" "$CLONE_DIR"
 
 cd "$CLONE_DIR"
+
 echo "Building Docker image $DOCKER_REPO..."
 docker build -t "$DOCKER_REPO" .
 
 echo "Logging in to Docker Hub..."
-echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin
+echo "$DOCKER_PWD" | docker login -u "$DOCKER_USER" --password-stdin
 
 echo "Pushing $DOCKER_REPO to Docker Hub..."
 docker push "$DOCKER_REPO"
@@ -34,5 +37,4 @@ docker push "$DOCKER_REPO"
 cd ..
 rm -rf "$CLONE_DIR"
 
-echo "Image succesfully pushed to Docker Hub: $DOCKER_REPO"
-
+echo "✅ Image successfully pushed to Docker Hub: $DOCKER_REPO"
